supported --float_type

+++ csubst/combination.py
@@ -48,7 +48,7 @@ def get_node_combinations(g, target_nodes=None, arity=2, check_attr=None, verbos
         axis = (len(index_combinations), arity)
         mmap_out = os.path.join(os.getcwd(), 'tmp.csubst.node_combinations.mmap')
         if os.path.exists(mmap_out): os.unlink(mmap_out)
-        df_mmap = numpy.memmap(mmap_out, dtype=numpy.int64, shape=axis, mode='w+')
+        df_mmap = numpy.memmap(mmap_out, dtype=numpy.int32, shape=axis, mode='w+')
         chunks,starts = get_chunks(index_combinations, g['nslots'])
         joblib.Parallel(n_jobs=g['nslots'], max_nbytes=None, backend='multiprocessing')(
             joblib.delayed(node_union)
